grids_follow_player
creates a 3 by 3 grid that follows the player, remembering the grids that are currently loaded until they are deleted. The step when they are deleted in the future can be storing it on the server, moving into a grid of an ID can then be fetched.... Black dots are created at the initial grids randomly, this is the beginning of the voronoi grids for future releases. Thank you